+++ src/components/random-content/random-content.styles.ts
@@ -1,0 +1,11 @@
+import { css } from 'lit';
+
+export default css`
+  :host {
+    display: contents;
+  }
+
+  ::slotted(*:not([data-visible])) {
+    display: none !important;
+  }
+`;